requirements.txt
# Core
streamlit>=1.32.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# Data Sources
alpaca-py>=0.13.0
finnhub-python>=2.4.0
twelvedata>=1.2.0
python-binance>=1.0.19
websocket-client>=1.7.0
websockets>=12.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Technical Analysis
pandas-ta>=0.3.14b0

# Visualization
plotly>=5.18.0

# LLM Integration
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0

# Storage
duckdb>=0.10.0
sqlalchemy>=2.0.0

# Utilities
httpx>=0.27.0
tenacity>=8.2.0
structlog>=24.1.0
requests>=2.31.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.0
